!3246 fix(#I5QTY6): slider can't drag in dialog * chore: bump version 6.10.3-beta02 * test: 更新单元测试 * refactor: 更新示例文档 * refactor: 消除警告信息 * refactor: 增加最大值最小值设置 值更改为 double 数据类型 * refactor: 更新脚本 * fix: 修复 Slider 在弹窗中无法拖动问题

+++ src/BootstrapBlazor.Shared/Samples/Sliders.razor.cs
@@ -11,7 +11,7 @@ namespace BootstrapBlazor.Shared.Samples;
 /// </summary>
 public sealed partial class Sliders
 {
-    private int BindValue { get; set; } = 50;
+    private double BindValue { get; set; } = 50;
 
     /// <summary>
     /// 获得属性方法
